avoid using Weld-specific API to obtain current injection point
It turns out that when calling the synthetic bean creation function,
it is possible to obtain the "current" injection point simply
by looking up the bean of type `InjectionPoint`. There's no reason
to use the Weld-specific API.

This means that there's just 1 tiny piece of Weld-specific code:
the method to set a synthetic bean priority. This needs to be added
to the CDI API.